requirements.txt
numpy
torch
librosa
matplotlib
scipy
opencv-python
tqdm
Pillow
